+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # OpenCore Legacy Patcher changelog
 
+## 0.5.2
+
 ## 0.5.1
 - Add support for `APPLE SSD TS0128F/256F` SSDs in macOS Ventura
   - ie. stock SSD found in MacBookAir6,x